% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curated_motifs}
\alias{curated_motifs}
\title{Curated DNA Binding Motifs Collection}
\format{
A \code{\link[universalmotif]{universalMotif-list}} object containing 3,255 motifs.
}
\source{
Curated from multiple public databases
}
\usage{
curated_motifs
}
\description{
A collection of 3,255 high-quality, manually curated non-redundant DNA binding motifs
corresponding to 1,410 mouse/human genes.
}
\keyword{datasets}
