% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curated_motifs_meta}
\alias{curated_motifs_meta}
\title{Metadata for Curated Motifs}
\format{
A data frame with metadata information for each motif.
}
\usage{
curated_motifs_meta
}
\description{
Detailed metadata for the 3,255 curated DNA binding motifs.
}
\keyword{datasets}
