% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_rawVScurated}
\alias{plot_rawVScurated}
\title{Plot comparison between raw and curated motifs for a specific gene}
\usage{
plot_rawVScurated(gene = NULL)
}
\arguments{
\item{gene}{Gene symbol to plot motifs for}
}
\value{
A ggplot object showing motif comparisons
}
\description{
Plot comparison between raw and curated motifs for a specific gene
}
