% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_motifs}
\alias{raw_motifs}
\title{Raw DNA Binding Motifs Collection}
\format{
A \code{\link[universalmotif]{universalMotif-list}} object.
}
\usage{
raw_motifs
}
\description{
The original, unprocessed collection of DNA binding motifs before curation.
}
\keyword{datasets}
