% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_motifs_meta}
\alias{raw_motifs_meta}
\title{Metadata for Raw Motifs}
\format{
A data frame with metadata information.
}
\usage{
raw_motifs_meta
}
\description{
Metadata for the raw, unprocessed DNA binding motifs.
}
\keyword{datasets}
